+++ lib/cli/deploy_builder.rs
@@ -145,7 +145,7 @@ impl<'a> DeployBuilder<'a> {
 
         let account = match initiator_addr_and_secret_key.initiator_addr() {
             InitiatorAddr::PublicKey(public_key) => public_key,
-            InitiatorAddr::AccountHash(_) => unreachable!(),
+            InitiatorAddr::AccountHash(_) | InitiatorAddr::Eoa(_) => unreachable!(),
         };
 
         let dependencies = dependencies.into_iter().unique().collect();